% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_operations_combine_addtopic_subsetmir.R
\name{add_col_topic}
\alias{add_col_topic}
\title{Add topic column to data frame}
\usage{
add_col_topic(df, col.topic = "Topic", topic.name = "Topic1")
}
\arguments{
\item{df}{Data frame which the topic column is added to.}

\item{col.topic}{String. Name of the topic column to be created.}

\item{topic.name}{String. Topic name to be contained in \code{col.topic}.}
}
\value{
Data frame with a topic column added.
}
\description{
Add topic column to a data frame.
}
\details{
Add a topic column to a data frame. This topic column is named \code{col.topic} and
contains the string \code{topic.name}.
}
\seealso{
\code{\link[=assign_topic]{assign_topic()}}
}
