% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals_score.R
\docType{data}
\name{animal_keywords}
\alias{animal_keywords}
\title{Keywords - animals.}
\format{
An object of class \code{character} of length 12.
}
\usage{
animal_keywords
}
\description{
Keywords to identify abstracts using animal models.
}
\keyword{datasets}
