% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_score.R
\docType{data}
\name{biomarker_keywords}
\alias{biomarker_keywords}
\title{Keywords - biomarkers.}
\format{
An object of class \code{character} of length 18.
}
\usage{
biomarker_keywords
}
\description{
Keywords to identify abstracts reporting about miRNAs as biomarkers.
}
\keyword{datasets}
