% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_operations_combine_addtopic_subsetmir.R
\name{combine_df}
\alias{combine_df}
\title{Combine data frames into one data frame}
\usage{
combine_df(...)
}
\arguments{
\item{...}{Data frames to combine into one data frame.
Data frames must have the same number of columns and the same column names.}
}
\value{
Combined data frame.
}
\description{
Combine data frames into one data frame.
}
\details{
Combine data frames into one data frame. \code{combine_df()}
accepts several data frames that are combined into one data frame.
Data frames to be combined must have the same number
of columns and the same column names.
}
\seealso{
Other combine functions: 
\code{\link{combine_mir}()}
}
\concept{combine functions}
