% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mir.R
\name{combine_mir}
\alias{combine_mir}
\title{Combine miRNA vectors into one}
\usage{
combine_mir(...)
}
\arguments{
\item{...}{Character vectors. Character vectors containing miRNA names.}
}
\value{
Combined character vector containing miRNA names.
}
\description{
Combine miRNA vectors into one.
}
\details{
Combine miRNA vectors into one. miRNA names occurring more than once
are reduced to one instance.
}
\seealso{
\code{\link[=get_mir]{get_mir()}}

Other combine functions: 
\code{\link{combine_df}()}
}
\concept{combine functions}
