% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stopwords.R
\name{combine_stopwords}
\alias{combine_stopwords}
\title{Combine data frames containing stop words}
\usage{
combine_stopwords(...)
}
\arguments{
\item{...}{Data frames with stop words. Data frames must have two columns
named "word" and "lexicon".}
}
\value{
Combined data frame with stop words.
}
\description{
Combine data frames containing stop words into one data frame.
}
\details{
Combine data frames containing stop words into one data frame. Provided
data frames must have two columns named "word" and "lexicon".
}
\seealso{
\code{\link[=generate_stopwords]{generate_stopwords()}}, \link{stopwords_miretrieve}, \link[tidytext:stop_words]{tidytext::stop_words}

Other stopword functions: 
\code{\link{generate_stopwords}()}
}
\concept{stopword functions}
