% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mir_single_topic.R
\name{count_mir}
\alias{count_mir}
\title{Count miRNA names in a data frame}
\usage{
count_mir(df, col.mir = miRNA)
}
\arguments{
\item{df}{Data frame containing miRNA names.}

\item{col.mir}{Symbol. Column containing miRNA names.}
}
\value{
Data frame. Data frame containing miRNA names and their
respective frequency.
}
\description{
Count occurrence of miRNA names in a data frame.
}
\details{
Count occurrence of miRNA names in  a data frame. The count of miRNA names
is returned as a separate data frame, only listing the miRNA
names and their respective frequency.
}
\seealso{
\code{\link[=plot_mir_count]{plot_mir_count()}}, \code{\link[=count_mir_threshold]{count_mir_threshold()}}, \code{\link[=plot_mir_count_threshold]{plot_mir_count_threshold()}}

Other count functions: 
\code{\link{count_mir_threshold}()},
\code{\link{count_snp}()},
\code{\link{plot_mir_count_threshold}()},
\code{\link{plot_mir_count}()}
}
\concept{count functions}
