% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymorphism.R
\name{count_snp}
\alias{count_snp}
\title{Count SNPs in a data frame}
\usage{
count_snp(df, col.snp = SNPs, col.pmid = PMID)
}
\arguments{
\item{df}{Data frame containing SNPs and PubMed IDs.}

\item{col.snp}{Symbol. Column containing SNPs.}

\item{col.pmid}{Symbol. Column containing PubMed IDs.}
}
\value{
Data frame. Data frame containing SNPs and their
respective frequency.
}
\description{
Count occurrence of SNPs in a data frame.
}
\details{
Count occurrence of SNPs in  a data frame. The count of SNPs
is returned as a separate data frame, only listing the SNPs
and their respective frequency.
}
\seealso{
\code{\link[=extract_snp]{extract_snp()}},
\code{\link[=get_snp]{get_snp()}},
\code{\link[=subset_snp]{subset_snp()}}

Other count functions: 
\code{\link{count_mir_threshold}()},
\code{\link{count_mir}()},
\code{\link{plot_mir_count_threshold}()},
\code{\link{plot_mir_count}()}
}
\concept{count functions}
