% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_crc.R
\docType{data}
\name{df_crc}
\alias{df_crc}
\title{Dataset of PubMed data of miRNAs in Colorectal Cancer}
\format{
A data frame.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/}
}
\usage{
df_crc
}
\description{
A dataset PubMed abstracts of miRNAs in Colorectal Cancer.
}
\keyword{datasets}
