% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_diabetes.R
\docType{data}
\name{df_diabetes}
\alias{df_diabetes}
\title{Dataset of PubMed data of miRNAs in diabetes}
\format{A data frame.}
\source{
\url{http://pubmed.org/}
}
\usage{
df_diabetes
}
\description{
A dataset PubMed abstracts of miRNAs in diabetes. miRNAs were extracted and
first 500 rows were kept to minimize storage. Dataset is used for exemplary
purposes in the vignette only.
}
\keyword{datasets}
