% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mirtarbase.R
\docType{data}
\name{df_mirtarbase}
\alias{df_mirtarbase}
\title{miRTarBase version 8.0}
\format{
A data frame with the columns "miRNA_tarbase", "Target", and "PMID".
}
\source{
\url{https://miRTarBase.cuhk.edu.cn:443/}
}
\usage{
df_mirtarbase
}
\description{
The most recent miRTarBase version 8.0, containing miRNA stem,
capitalized targets, and PMIDs.
}
\details{
miRTarBase was published in

Hsi-Yuan Huang, Yang-Chi-Dung Lin, Jing Li, et al.,
miRTarBase 2020: updates to the experimentally validated microRNA–target
interaction database, Nucleic Acids Research, Volume 48, Issue D1,
08 January 2020, Pages D148–D154, https://doi.org/10.1093/nar/gkz896
}
\keyword{datasets}
