% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_panc.R
\docType{data}
\name{df_panc}
\alias{df_panc}
\title{Dataset of PubMed data of miRNAs in Pancreatic Cancer}
\format{
A data frame.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/}
}
\usage{
df_panc
}
\description{
A dataset PubMed abstracts of miRNAs in Pancreatic Cancer.
}
\keyword{datasets}
