% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_targets.R
\docType{data}
\name{df_targets}
\alias{df_targets}
\title{Merged Dataset of miRNA and their targets}
\format{A data frame.}
\source{
\url{http://pubmed.org/}
}
\usage{
df_targets
}
\description{
Merged Dataset of miRNA and their targets in Pancreatic Cancer,
Colorectal Cancer and their respective targets from MirTARBase.
}
\keyword{datasets}
