% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_test.R
\docType{data}
\name{df_test}
\alias{df_test}
\title{Test dataset of PubMed abstracts}
\format{
A data frame.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/}
}
\usage{
df_test
}
\description{
Test dataset of 20 PubMed abstracts.
}
\keyword{datasets}
