% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_shared_mir.R
\name{get_shared_mir_vec}
\alias{get_shared_mir_vec}
\title{Get miRNA names in common between two vectors}
\usage{
get_shared_mir_vec(mirna.vec.1, mirna.vec.2)
}
\arguments{
\item{mirna.vec.1}{Character vector. First vector containing miRNA
names.}

\item{mirna.vec.2}{Character vector. Second vector containing miRNA
names.}
}
\value{
Character vector containing miRNA names in common between two
vectors.
}
\description{
Get miRNA names in common between two vectors.
}
\details{
Get miRNA names in common between two vectors.
\code{get_shared_mir_vec()} compares two vectors containing miRNA names and
returns the miRNA names that are in both vectors.
}
\seealso{
Other get functions: 
\code{\link{get_distinct_mir_df}()},
\code{\link{get_distinct_mir_vec}()},
\code{\link{get_mir}()},
\code{\link{get_pmid}()},
\code{\link{get_shared_mir_df}()},
\code{\link{get_snp}()}
}
\concept{get functions}
