% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_ngram_roxygen.R
\docType{data}
\name{ngram_stopwords}
\alias{ngram_stopwords}
\title{Stop words for n-grams}
\format{
Character vector.
}
\source{
tidytext::stop_words
}
\usage{
ngram_stopwords
}
\description{
Vector containing stop words for n-grams, based on
tidytext::stop_words.
}
\keyword{datasets}
