% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients_score.R
\docType{data}
\name{patients_keywords}
\alias{patients_keywords}
\title{Keywords - patients.}
\format{
An object of class \code{character} of length 10.
}
\usage{
patients_keywords
}
\description{
Keywords to identify abstracts investigating miRNAs in patients.
}
\keyword{datasets}
