% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_save_files.R
\name{read_pubmed_xml}
\alias{read_pubmed_xml}
\title{Convert xml-file from PubMed into a data frame}
\usage{
read_pubmed_xml(xml_file, topic = NULL)
}
\arguments{
\item{xml_file}{xml-file, downloaded from PubMed.}

\item{topic}{String. Optional. If provided, adds a "Topic" column containing
\code{topic}.}
}
\value{
Data frame containing PubMed-IDs, abstracts, abstract titles,
publication years, languages, and article types.
}
\description{
Convert xml-file from PubMed into a data frame.
}
\details{
Converts an xml-file from PubMed into a data
frame.
The xml-file should contain PubMed-IDs, abstracts from research articles,
abstract title, publication year, abstract language, and article type.
The data frame created holds at least six columns, namely
\itemize{
\item \code{PMID}, containing the PubMed-ID,
\item \code{Year}, containing the publication year,
\item \code{Title}, containing the title of the abstracts,
\item \code{Abstract}, containing the actual abstract,
\item \code{Language}, containing the language(s) of the paper,
\item \code{Type}, containing the article type.
}

If \code{topic} is provided, a "Topic" column is added, assigning all abstracts in
\code{df} to \code{topic}.

\code{read_pubmed_medline()} is faster than \code{read_pubmed_xml()} and thus
recommended.
}
\seealso{
\code{\link[=read_pubmed_medline]{read_pubmed_medline()}}

Other external data functions: \code{\link{read_pubmed_medline}},
  \code{\link{save_excel}}, \code{\link{save_plot}}
}
\concept{external data functions}
