% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_save_files.R
\name{save_excel}
\alias{save_excel}
\title{Save data frame(s) as xlsx-file}
\usage{
save_excel(..., excel_file = "miRetrieve_data.xlsx")
}
\arguments{
\item{...}{Data frame(s) to save.}

\item{excel_file}{String. File name that \code{...} shall
be saved to. Must end in ".xlsx".}
}
\value{
xlsx-file, locally saved.
}
\description{
Save data frame(s) locally as an xlsx-file.
}
\details{
Saves data frame locally as an xlsx-file. If more than
one data frame is provided, data frames are saved
in an xlsx-file with one sheet per data frame.

Wrapper function of \code{write.xlsx()} from \pkg{openxlsx}.
}
\seealso{
\code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx()}}

Other external data functions: 
\code{\link{read_pubmed_jats}()},
\code{\link{read_pubmed}()},
\code{\link{save_plot}()}
}
\concept{external data functions}
