% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_doc.R
\docType{data}
\name{stopwords_miretrieve}
\alias{stopwords_miretrieve}
\title{Stop words for text mining with miRetrieve}
\format{
Tibble.
\itemize{
\item \code{word}: Column containing stop words. Pulled from various PubMed
abstracts.
\item \code{lexicon}: Column specifying lexicon.
}
}
\source{
tidytext::stop_words; manually created from various PubMed abstracts.
}
\usage{
stopwords_miretrieve
}
\description{
Data frame containing English stop words, PubMed stop words,
and common 2-gram stopwords. English stop words are based on
tidytext::stop_words, while PubMed stop words are manually curated from
PubMed abstracts
}
\keyword{datasets}
