% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_doc.R
\docType{data}
\name{stopwords_pubmed}
\alias{stopwords_pubmed}
\title{Stop words for text mining from PubMed abstracts}
\format{
Tibble.
\itemize{
\item \code{word}: Column containing stop words. Pulled from various PubMed
abstracts.
\item \code{lexicon}: Column specifying lexicon.
}
}
\source{
Manually created from various PubMed abstracts.
}
\usage{
stopwords_pubmed
}
\description{
Data frame containing PubMed stop words, manually curated from
PubMed abstracts
}
\keyword{datasets}
