% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_articles_review_year.R
\name{subset_research}
\alias{subset_research}
\title{Subset data frame for abstracts of research articles}
\usage{
subset_research(df, col.type = Type)
}
\arguments{
\item{df}{Data frame containing article types.}

\item{col.type}{Symbol. Column containing articles types.}
}
\value{
Data frame containing abstracts of research articles only.
}
\description{
Subset data frame for abstracts of research articles only.
}
\details{
Subset data frame for abstracts of research articles only.
At the same time, abstracts from other article types such as \emph{Review},
\emph{Letter}, etc. are dropped.
}
\seealso{
\code{\link[=subset_review]{subset_review()}}, \code{\link[=subset_year]{subset_year()}}

Other subset functions: 
\code{\link{subset_df}()},
\code{\link{subset_mir_threshold}()},
\code{\link{subset_mir}()},
\code{\link{subset_review}()},
\code{\link{subset_snp}()},
\code{\link{subset_year}()}
}
\concept{subset functions}
